Solution to Check Profit or Loss

+++ Check Profit or Loss@@6788dd63490637cfc44e1b98/code.c
@@ -5,7 +5,7 @@ int main() {
     scanf("%d",&a);
     if ((int)a>a)
     printf("Profit");
-    else if(a<a)
+    else if((int)a<a)
     printf("Loss");
     else
     printf("No Profit No Loss")
